🐛 Fix BlockArgumentParser is not a constructor
