+++ src/parsers/EntityArgumentParser.ts
@@ -1,6 +1,4 @@
-import { getNbtdocRegistryId } from '../CommandTree'
 import { locale } from '../locales'
-import { getCompletions, getSafeCategory } from '../types/ClientCache'
 import { NodeRange } from '../nodes/ArgumentNode'
 import { EntityNode } from '../nodes/EntityNode'
 import { IdentityNode } from '../nodes/IdentityNode'
@@ -9,6 +7,7 @@ import { NumberNode } from '../nodes/NumberNode'
 import { NumberRangeNode } from '../nodes/NumberRangeNode'
 import { EntitySelectorNodeChars, SelectorAdvancementsNode, SelectorArgumentKey, SelectorArgumentKeys, SelectorArgumentNodeChars, SelectorArgumentsNode, SelectorCriteriaNode, SelectorScoresNode, SelectorSortMethod } from '../nodes/SelectorArgumentsNode'
 import { StringNode } from '../nodes/StringNode'
+import { getCompletions, getSafeCategory } from '../types/ClientCache'
 import { ArgumentParserResult, combineArgumentParserResult } from '../types/Parser'
 import { ParsingContext } from '../types/ParsingContext'
 import { ErrorCode, ParsingError } from '../types/ParsingError'
@@ -412,4 +411,15 @@ export class EntityArgumentParser extends ArgumentParser<EntityNode> {
     getExamples(): string[] {
         return ['Player', '0123', '@e', '@e[type=foo]', 'dd12be42-52a9-4a91-a8a1-11c01849e498']
     }
+}
+
+export function getNbtdocRegistryId(entity: EntityNode): null | string {
+    if (entity.variable === 'a' || entity.variable === 'p' || entity.variable === 'r') {
+        return 'minecraft:player'
+    }
+    const firstID = entity.argument.type?.[0]
+    if (firstID && !firstID.isTag) {
+        return firstID.toString()
+    }
+    return null
 }
