+++ src/test/parsers/EntityArgumentParser.spec.ts
@@ -2,7 +2,7 @@ import assert = require('power-assert')
 import { describe, it } from 'mocha'
 import { CompletionItemKind, DiagnosticSeverity } from 'vscode-languageserver'
 import { ArgumentParserManager } from '../../parsers/ArgumentParserManager'
-import { EntityArgumentParser } from '../../parsers/EntityArgumentParser'
+import { EntityArgumentParser, getNbtdocRegistryId } from '../../parsers/EntityArgumentParser'
 import { constructConfig } from '../../types/Config'
 import { NodeRange } from '../../nodes/ArgumentNode'
 import { EntityNode } from '../../nodes/EntityNode'
@@ -643,6 +643,42 @@ describe('EntityArgumentParser Tests', () => {
                 const actual = parser.parse(new StringReader('@a[tag=foo,tag=!bar]'), ctx)
                 assert.deepStrictEqual(actual.errors, [])
             })
+        })
+    })
+    describe('getNbtdocRegistryId() Tests', () => {
+        it('Should return the respective id', () => {
+            const id = new IdentityNode('minecraft', ['spgoding'])
+            const argument = new SelectorArgumentsNode()
+            argument.type = [id]
+            const entity = new EntityNode(undefined, 'e', argument)
+            const actual = getNbtdocRegistryId(entity)
+            assert(actual === 'minecraft:spgoding')
+        })
+        it('Should return minecraft:player for @a', () => {
+            const entity = new EntityNode(undefined, 'a', new SelectorArgumentsNode())
+            const actual = getNbtdocRegistryId(entity)
+            assert(actual === 'minecraft:player')
+        })
+        it('Should return null when there is no type', () => {
+            const argument = new SelectorArgumentsNode()
+            const entity = new EntityNode(undefined, 'e', argument)
+            const actual = getNbtdocRegistryId(entity)
+            assert(actual === null)
+        })
+        it('Should return null when the type is empty', () => {
+            const argument = new SelectorArgumentsNode()
+            argument.type = []
+            const entity = new EntityNode(undefined, 'e', argument)
+            const actual = getNbtdocRegistryId(entity)
+            assert(actual === null)
+        })
+        it('Should return null when the first type is a tag', () => {
+            const id = new IdentityNode('minecraft', ['spgoding'], true)
+            const argument = new SelectorArgumentsNode()
+            argument.type = [id]
+            const entity = new EntityNode(undefined, 'e', argument)
+            const actual = getNbtdocRegistryId(entity)
+            assert(actual === null)
         })
     })
 })
